Move common test code into own project
The previous approach of adding the class output
of the tests as an dependency was brittle and prevented
proper incremental builds when tests are compiled.
