+++ test-common/build.gradle.kts
@@ -1,0 +1,10 @@
+plugins {
+    `java-library`
+}
+
+dependencies {
+    api(testLibs.bundles.mockito)
+    api(testLibs.junit.api)
+    api(testLibs.junit.params)
+    runtimeOnly(testLibs.junit.engine)
+}
